fix(pyparticles): Cast render coordinates to int for Pygame 2.5 compatibility

+++ game_forge/pyparticles/src/pyparticles/rendering/canvas.py
@@ -128,8 +128,8 @@ class Canvas:
             if not np.any(mask): continue
             
             sprite = self.sprites[t]
-            xs = sx[mask]
-            ys = sy[mask]
+            xs = sx[mask].astype(int)
+            ys = sy[mask].astype(int)
             
             # Create (sprite, (x, y)) tuples
             # Zip is fast iterator
